actually builds on 1.4.2
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@908 406a0c4d-033a-0410-8de8-e80135713968

+++ api/java/test.java
@@ -130,7 +130,7 @@ public class test
 
 						case SphinxClient.SPH_ATTR_TIMESTAMP:
 							Long iStamp = (Long) info.attrValues.get(a);
-							Date date = new Date ( iStamp*1000 );
+							Date date = new Date ( iStamp.longValue()*1000 );
 							System.out.print ( date.toString() );
 							break;
